+++ applications/swarm/hive/Swarm.py
@@ -53,7 +53,7 @@ class Swarm:
             raise ValueError("Swarm: the city can't be empty")
 
         if self._mongo.getSwarmByIdentifier(self._identifier) == None:
-            self._mongo.insertSwarm(self._identifier, args.swarm_num_agent, args.user_email, self._name, self._host, self._g.swarm['swarm_turns'], self._g.swarm['swarm_cycles'], args.swarm_city)
+            self._mongo.insertSwarm(self._identifier, args.swarm_num_agent, args.user_email, self._name, self._helper.getTimeNow(), self._host, self._g.swarm['swarm_turns'], self._g.swarm['swarm_cycles'], args.swarm_city)
 
         self.syncFromDB()